Move tests to DelayedActionRendererTest
